+++ src/main/java/com/oneday/researcher/entity/Rsr_sub.java
@@ -14,10 +14,10 @@ import lombok.Setter;
 public class Rsr_sub {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long rsr_sub_seq;
+    private Integer rsr_sub_seq;
 
-    private Long rs_seq;
-    private Long rsi_seq;
+    private Integer rs_seq;
+    private Integer rsi_seq;
     private String rsi_no;
     private String rsr_sub_etc;
     private String rsr_sub_type3;
